requirements.txt
Flask
Flask-Admin
flask-security
flask-mongoengine
Flask-SQLAlchemy
pymongo
dnspython
bcrypt
python-dotenv

